refactor(vm): Refactor vm tests
