+++ src/vm/vm_tests.rs
@@ -1,1528 +1,37 @@
-#[allow(clippy::too_many_lines)]
-#[cfg(test)]
-mod tests {
+use crate::{
+    compiler::{code::Instructions, Compiler},
+    object::{test_utils::check_constants, Object},
+    parser::parse,
+    vm::VM,
+};
 
-    use std::collections::HashMap;
+pub(crate) struct VmTestCase {
+    pub(crate) input: String,
+    pub(crate) expected: Object,
+}
 
-    use crate::{
-        compiler::{code::Instructions, Compiler},
-        object::{test_utils::check_constants, Object},
-        parser::parse,
-        vm::VM,
-    };
+pub(crate) fn run_vm_tests(tests: Vec<VmTestCase>) {
+    for test in tests {
+        println!("Running test: {}", test.input);
+        let program = parse(&test.input);
+        let mut compiler = Compiler::new();
+        compiler.compile(program).unwrap();
+        let bytecode = compiler.bytecode();
 
-    struct VmTestCase {
-        input: String,
-        expected: Object,
-    }
-
-    fn run_vm_tests(tests: Vec<VmTestCase>) {
-        for test in tests {
-            println!("Running test: {}", test.input);
-            let program = parse(&test.input);
-            let mut compiler = Compiler::new();
-            compiler.compile(program).unwrap();
-            let bytecode = compiler.bytecode();
-
-            for (i, constant) in bytecode.constants.iter().enumerate() {
-                match constant {
-                    Object::COMPILEDFUNCTION(cf) => {
-                        println!("Compiled function:");
-                        let instructions = Instructions::new(cf.instructions.clone());
-                        println!("{}", instructions);
-                    }
-                    _ => {
-                        println!("{}: {}", i, constant);
-                    }
+        for (i, constant) in bytecode.constants.iter().enumerate() {
+            match constant {
+                Object::COMPILEDFUNCTION(cf) => {
+                    println!("Compiled function:");
+                    let instructions = Instructions::new(cf.instructions.clone());
+                    println!("{}", instructions);
                 }
-            }
-
-            let mut vm = VM::new(bytecode);
-            vm.run().unwrap();
-            let got = vm.last_popped_stack_element().unwrap();
-            check_constants(&vec![test.expected], &vec![got]);
-        }
-    }
-
-    #[test]
-    fn test_integer_arithmetic() {
-        let tests = vec![
-            VmTestCase {
-                input: "1".to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: "2".to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: "1 + 2".to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: "1 - 2".to_string(),
-                expected: Object::INTEGER(-1),
-            },
-            VmTestCase {
-                input: "1 * 2".to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: "4 / 2".to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: "50 / 2 * 2 + 10 - 5".to_string(),
-                expected: Object::INTEGER(55),
-            },
-            VmTestCase {
-                input: "5 + 5 + 5 + 5 - 10".to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: "2 * 2 * 2 * 2 * 2".to_string(),
-                expected: Object::INTEGER(32),
-            },
-            VmTestCase {
-                input: "5 * 2 + 10".to_string(),
-                expected: Object::INTEGER(20),
-            },
-            VmTestCase {
-                input: "-1".to_string(),
-                expected: Object::INTEGER(-1),
-            },
-            VmTestCase {
-                input: "-10".to_string(),
-                expected: Object::INTEGER(-10),
-            },
-            VmTestCase {
-                input: "-50 + 100 + -50".to_string(),
-                expected: Object::INTEGER(0),
-            },
-            VmTestCase {
-                input: "(5 + 10 * 2 + 15 / 3) * 2 + -10".to_string(),
-                expected: Object::INTEGER(50),
-            },
-        ];
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_boolean_logic() {
-        let tests = vec![
-            VmTestCase {
-                input: "true".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "false".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "1 < 2".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "1 <= 2".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "2 <= 2".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "1 > 2".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "1 >= 2".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "2 >= 2".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "1 == 2".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "1 != 2".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "true == true".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "false == false".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "true == false".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "true && true".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "true && false".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "false && true".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "false && false".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "true || true".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "true || false".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "false || true".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "false || false".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "!true".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "!false".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "!5".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "!!true".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "!!false".to_string(),
-                expected: Object::BOOLEAN(false),
-            },
-            VmTestCase {
-                input: "!!5".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-            VmTestCase {
-                input: "!(if (false) { 5 })".to_string(),
-                expected: Object::BOOLEAN(true),
-            },
-        ];
-        run_vm_tests(tests);
-    }
-    #[test]
-    fn test_conditionals() {
-        let tests = vec![
-            VmTestCase {
-                input: "if (true) { 10 }".to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: "if (true) { 10 } else { 20 }".to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: "if (false) { 10 } else { 20 } ".to_string(),
-                expected: Object::INTEGER(20),
-            },
-            VmTestCase {
-                input: "if (1) { 10 }".to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: "if (1 < 2) { 10 }".to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: "if (1 < 2) { 10 } else { 20 }".to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: "if (1 > 2) { 10 } else { 20 }".to_string(),
-                expected: Object::INTEGER(20),
-            },
-            VmTestCase {
-                input: "if (1 > 2) { 10 }".to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: "if (false) { 10 }".to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: "if ((if (false) { 10 })) { 10 } else { 20 }".to_string(),
-                expected: Object::INTEGER(20),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_global_let_statements() {
-        let tests = vec![
-            VmTestCase {
-                input: "let one = 1; one".to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: "let one = 1; let two = 2; one + two".to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: "let one = 1; let two = one + one; one + two".to_string(),
-                expected: Object::INTEGER(3),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_string_expressions() {
-        let tests = vec![
-            VmTestCase {
-                input: "\"monkey\"".to_string(),
-                expected: Object::STRING("monkey".to_string()),
-            },
-            VmTestCase {
-                input: "\"mon\" + \"key\"".to_string(),
-                expected: Object::STRING("monkey".to_string()),
-            },
-            VmTestCase {
-                input: "\"mon\" + \"key\" + \"banana\"".to_string(),
-                expected: Object::STRING("monkeybanana".to_string()),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_array_expressions() {
-        let tests = vec![
-            VmTestCase {
-                input: "[]".to_string(),
-                expected: Object::ARRAY(vec![]),
-            },
-            VmTestCase {
-                input: "[1, 2, 3]".to_string(),
-                expected: Object::ARRAY(vec![
-                    Object::INTEGER(1),
-                    Object::INTEGER(2),
-                    Object::INTEGER(3),
-                ]),
-            },
-            VmTestCase {
-                input: "[1 + 2, 3 * 4, 5 + 6]".to_string(),
-                expected: Object::ARRAY(vec![
-                    Object::INTEGER(3),
-                    Object::INTEGER(12),
-                    Object::INTEGER(11),
-                ]),
-            },
-            VmTestCase {
-                input: "[\"yes\", false, [1,2]]".to_string(),
-                expected: Object::ARRAY(vec![
-                    Object::STRING("yes".to_string()),
-                    Object::BOOLEAN(false),
-                    Object::ARRAY(vec![Object::INTEGER(1), Object::INTEGER(2)]),
-                ]),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-    #[test]
-    fn test_hashmap_expressions() {
-        let tests = vec![
-            VmTestCase {
-                input: "{}".to_string(),
-                expected: Object::HASHMAP(HashMap::new()),
-            },
-            VmTestCase {
-                input: "{1:2, 2:3}".to_string(),
-                expected: Object::HASHMAP(
-                    vec![
-                        (Object::INTEGER(1), Object::INTEGER(2)),
-                        (Object::INTEGER(2), Object::INTEGER(3)),
-                    ]
-                    .into_iter()
-                    .collect(),
-                ),
-            },
-            VmTestCase {
-                input: "{1+1:2, 2*2:3}".to_string(),
-                expected: Object::HASHMAP(
-                    vec![
-                        (Object::INTEGER(2), Object::INTEGER(2)),
-                        (Object::INTEGER(4), Object::INTEGER(3)),
-                    ]
-                    .into_iter()
-                    .collect(),
-                ),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_index_expression() {
-        let tests = vec![
-            VmTestCase {
-                input: "[1, 2, 3][1]".to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: "[1, 2, 3][0 + 2]".to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: "[[1, 1, 1]][0][0]".to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: "[][0]".to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: "[1, 2, 3][99]".to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: "[1][-1]".to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: "{1: 1, 2: 2}[1]".to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: "{1: 1, 2: 2}[2]".to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: "{1: 1}[0]".to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: "{}[0]".to_string(),
-                expected: Object::NULL,
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_calling_functions_without_arguments() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let fivePlusTen = fn() { 5 + 10; };
-                    fivePlusTen();"#
-                    .to_string(),
-                expected: Object::INTEGER(15),
-            },
-            VmTestCase {
-                input: r#"
-                    let one = fn() { 1; };
-                    let two = fn() { 2; };
-                    one() + two()"#
-                    .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = fn() { 1 };
-                    let b = fn() { a() + 1 };
-                    let c = fn() { b() + 1 };
-                    c();"#
-                    .to_string(),
-                expected: Object::INTEGER(3),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_functions_with_return_statements() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let earlyExit = fn() { return 99; 100; };
-                    earlyExit();"#
-                    .to_string(),
-                expected: Object::INTEGER(99),
-            },
-            VmTestCase {
-                input: r#"
-                    let earlyExit = fn() { return 99; return 100; };
-                    earlyExit();"#
-                    .to_string(),
-                expected: Object::INTEGER(99),
-            },
-        ];
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_functions_without_return_value() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let noReturn = fn() { };
-                    noReturn();"#
-                    .to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: r#"
-                    let noReturn = fn() { };
-                    let noReturnTwo = fn() { noReturn(); };
-                    noReturn();
-                    noReturnTwo();"#
-                    .to_string(),
-                expected: Object::NULL,
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_first_class_functions() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                let returnsOne = fn() { 1; };
-                let returnsOneReturner = fn() { returnsOne; };
-                returnsOneReturner()();"#
-                    .to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"
-                    let returnsOneReturner = fn() {
-                        let returnsOne = fn() { 1; };
-                        returnsOne;
-                    };
-                    returnsOneReturner()();"#
-                    .to_string(),
-                expected: Object::INTEGER(1),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_calling_function_with_bindings() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let one = fn() { let one = 1; one };
-                    one();"#
-                    .to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"
-                    let oneAndTwo = fn() { let one = 1; let two = 2; one + two; };
-                    oneAndTwo();"#
-                    .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"
-                    let oneAndTwo = fn() { let one = 1; let two = 2; one + two; };
-                    let threeAndFour = fn() { let three = 3; let four = 4; three + four; };
-                    oneAndTwo() + threeAndFour();"#
-                    .to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: r#"
-                    let firstFoobar = fn() { let foobar = 50; foobar; };
-                    let secondFoobar = fn() { let foobar = 100; foobar; };
-                    firstFoobar() + secondFoobar();"#
-                    .to_string(),
-                expected: Object::INTEGER(150),
-            },
-            VmTestCase {
-                input: r#"
-                    let globalSeed = 50;
-                    let minusOne = fn() {
-                        let num = 1;
-                        globalSeed - num;
-                    }
-                    let minusTwo = fn() {
-                        let num = 2;
-                        globalSeed - num;
-                    }
-                    minusOne() + minusTwo();"#
-                    .to_string(),
-                expected: Object::INTEGER(97),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_calling_functions_with_arguments_and_bindings() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let identity = fn(a) { a; };
-                    identity(4);"#
-                    .to_string(),
-                expected: Object::INTEGER(4),
-            },
-            VmTestCase {
-                input: r#"
-                    let sum = fn(a, b) { a + b; };
-                    sum(1, 2);"#
-                    .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"
-                    let sum = fn(a, b) {
-                        let c = a + b;
-                        c;
-                    };
-                    sum(1, 2);"#
-                    .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"
-                    let sum = fn(a, b) {
-                        let c = a + b;
-                        c;
-                    };
-                    sum(1, 2) + sum(3, 4);"#
-                    .to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: r#"
-                    let sum = fn(a, b) {
-                        let c = a + b;
-                        c;
-                    };
-                    let outer = fn() {
-                        sum(1, 2) + sum(3, 4);
-                    };
-                    outer();"#
-                    .to_string(),
-                expected: Object::INTEGER(10),
-            },
-            VmTestCase {
-                input: r#"
-                    let globalNum = 10;
-                    let sum = fn(a, b) {
-                        let c = a + b;
-                        c + globalNum;
-                    };
-                    let outer = fn() {
-                        sum(1, 2) + sum(3, 4) + globalNum;
-                    };
-                    outer() + globalNum;"#
-                    .to_string(),
-                expected: Object::INTEGER(50),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_calling_functions_with_wrong_arguments() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    fn() { 1; }(1);"#
-                    .to_string(),
-                expected: Object::ERROR("Wrong number of arguments: want=0, got=1".to_string()),
-            },
-            VmTestCase {
-                input: r#"
-                    fn(a) { a; }();"#
-                    .to_string(),
-                expected: Object::ERROR("Wrong number of arguments: want=1, got=0".to_string()),
-            },
-            VmTestCase {
-                input: r#"
-                    fn(a, b) { a + b; }(1);"#
-                    .to_string(),
-                expected: Object::ERROR("Wrong number of arguments: want=2, got=1".to_string()),
-            },
-        ];
-
-        for test in tests {
-            println!("Running test: {}", test.input);
-            let program = parse(&test.input);
-            let mut compiler = Compiler::new();
-            compiler.compile(program).unwrap();
-            let bytecode = compiler.bytecode();
-
-            let mut vm = VM::new(bytecode);
-            match vm.run() {
-                Ok(_) => {
-                    panic!("Expected error, but got no error");
-                }
-                Err(e) => match test.expected {
-                    Object::ERROR(msg) => {
-                        assert_eq!(e, msg);
-                    }
-                    _ => {
-                        unreachable!("Poorly written test, the expected value should be an error");
-                    }
-                },
+                _ => println!("{}: {}", i, constant),
             }
         }
-    }
 
-    #[test]
-    fn test_builtin_functions() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"len("")"#.to_string(),
-                expected: Object::INTEGER(0),
-            },
-            VmTestCase {
-                input: r#"len("four")"#.to_string(),
-                expected: Object::INTEGER(4),
-            },
-            VmTestCase {
-                input: r#"len("hello world")"#.to_string(),
-                expected: Object::INTEGER(11),
-            },
-            VmTestCase {
-                input: r#"len(1)"#.to_string(),
-                expected: Object::ERROR("argument to `len` not supported, got INTEGER".to_string()),
-            },
-            VmTestCase {
-                input: r#"len("one", "two")"#.to_string(),
-                expected: Object::ERROR("wrong number of arguments. got=2, want=1".to_string()),
-            },
-            VmTestCase {
-                input: r#"len([1, 2, 3])"#.to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"len([])"#.to_string(),
-                expected: Object::INTEGER(0),
-            },
-            VmTestCase {
-                input: r#"len([1, 2, 3], [4, 5, 6])"#.to_string(),
-                expected: Object::ERROR("wrong number of arguments. got=2, want=1".to_string()),
-            },
-            VmTestCase {
-                input: r#"first([1, 2, 3])"#.to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"first([])"#.to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: r#"first(1)"#.to_string(),
-                expected: Object::ERROR(
-                    "argument to `first` not supported, must be ARRAY, got INTEGER".to_string(),
-                ),
-            },
-            VmTestCase {
-                input: r#"last([1, 2, 3])"#.to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"last([])"#.to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: r#"last(1)"#.to_string(),
-                expected: Object::ERROR(
-                    "argument to `last` not supported, must be ARRAY, got INTEGER".to_string(),
-                ),
-            },
-            VmTestCase {
-                input: r#"rest([1, 2, 3])"#.to_string(),
-                expected: Object::ARRAY(vec![Object::INTEGER(2), Object::INTEGER(3)]),
-            },
-            VmTestCase {
-                input: r#"rest([])"#.to_string(),
-                expected: Object::NULL,
-            },
-            VmTestCase {
-                input: r#"push([], 1)"#.to_string(),
-                expected: Object::ARRAY(vec![Object::INTEGER(1)]),
-            },
-            VmTestCase {
-                input: r#"push(1, 1)"#.to_string(),
-                expected: Object::ERROR(
-                    "argument to `push` not supported, must be ARRAY, got INTEGER".to_string(),
-                ),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_closures() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                let newClosure = fn(a) {
-                    fn() { a; };
-                };
-                let closure = newClosure(99);
-                closure();"#
-                    .to_string(),
-                expected: Object::INTEGER(99),
-            },
-            VmTestCase {
-                input: r#"
-                let newAdder = fn(a, b) {
-                    fn(c) { a + b + c };
-                };
-                let adder = newAdder(1, 2);
-                adder(8);"#
-                    .to_string(),
-                expected: Object::INTEGER(11),
-            },
-            VmTestCase {
-                input: r#"
-                let newAdder = fn(a, b) {
-                    let c = a + b;
-                    fn(d) { c + d };
-                };
-                let adder = newAdder(1, 2);
-                adder(8);"#
-                    .to_string(),
-                expected: Object::INTEGER(11),
-            },
-            VmTestCase {
-                input: r#"
-                let newAdderOuter = fn(a, b) {
-                    let c = a + b;
-                    fn(d) {
-                        let e = d + c;
-                        fn(f) { e + f; };
-                    };
-                };
-                let newAdderInner = newAdderOuter(1, 2)
-                let adder = newAdderInner(3);
-                adder(8);"#
-                    .to_string(),
-                expected: Object::INTEGER(14),
-            },
-            VmTestCase {
-                input: r#"
-                let a = 1;
-                let newAdderOuter = fn(b) {
-                    fn(c) {
-                        fn(d) { a + b + c + d };
-                    };
-                };
-                let newAdderInner = newAdderOuter(2)
-                let adder = newAdderInner(3);
-                adder(8);"#
-                    .to_string(),
-                expected: Object::INTEGER(14),
-            },
-            VmTestCase {
-                input: r#"
-                let newClosure = fn(a, b) {
-                    let one = fn() { a; };
-                    let two = fn() { b; };
-                    fn() { one() + two(); };
-                };
-                let closure = newClosure(9, 90);
-                closure();"#
-                    .to_string(),
-                expected: Object::INTEGER(99),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_recursive_functions() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                let countDown = fn(x) {
-                    if (x == 0) {
-                        return 0;
-                    } else {
-                        countDown(x - 1);
-                    }
-                };
-                countDown(1);"#
-                    .to_string(),
-                expected: Object::INTEGER(0),
-            },
-            VmTestCase {
-                input: r#"
-                let countDown = fn(x) {
-                    if (x == 0) {
-                        return 0;
-                    } else {
-                        countDown(x - 1);
-                    }
-                };
-                let wrapper = fn() {
-                    countDown(1);
-                };
-                wrapper();"#
-                    .to_string(),
-                expected: Object::INTEGER(0),
-            },
-            VmTestCase {
-                input: r#"
-                let wrapper = fn() {
-                    let countDown = fn(x) {
-                        if (x == 0) {
-                            return 0;
-                        } else {
-                            countDown(x - 1);
-                        }
-                    };
-                    countDown(1);
-                };
-                wrapper();"#
-                    .to_string(),
-                expected: Object::INTEGER(0),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_recursive_fibonacci() {
-        let tests = vec![VmTestCase {
-            input: r#"
-                let fibonacci = fn(x) {
-                    if (x == 0) {
-                        return 0;
-                    } else {
-                        if (x == 1) {
-                            return 1;
-                        } else {
-                            fibonacci(x - 1) + fibonacci(x - 2);
-                        }
-                    }
-                };
-                fibonacci(15);"#
-                .to_string(),
-            expected: Object::INTEGER(610),
-        }];
-
-        run_vm_tests(tests);
-    }
-    #[test]
-    fn test_use_same_varaible_multiple_times() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    let b = a;
-                    let c = a + b + 1;
-                    c"#
-                .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    let b = a;
-                    let c = a + b + 1;
-                    let a = 2;
-                    let b = a;
-                    let c = a + b + 1;
-                    c"#
-                .to_string(),
-                expected: Object::INTEGER(5),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = "hello";
-                    let b = "world";
-                    let c = a + b;
-                    let d = a + b + c;
-                    d"#
-                .to_string(),
-                expected: Object::STRING("helloworldhelloworld".to_string()),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_array_multiple_ussage() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                let array = [1,2,3];
-                let value =  array[1] + array[2];
-                value"#
-                    .to_string(),
-                expected: Object::INTEGER(5),
-            },
-            VmTestCase {
-                input: r#"
-                let array = [1,2,3];
-                let array = push(array, 4);
-                array"#
-                    .to_string(),
-                expected: Object::ARRAY(vec![
-                    Object::INTEGER(1),
-                    Object::INTEGER(2),
-                    Object::INTEGER(3),
-                    Object::INTEGER(4),
-                ]),
-            },
-        ];
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_shadowing_same_type() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let x = 4;
-                    let x = 5;
-                    x"#
-                .to_string(),
-                expected: Object::INTEGER(5),
-            },
-            VmTestCase {
-                input: r#"
-                    let x = [1,2,3];
-                    let x = [4,5,6];
-                    x"#
-                .to_string(),
-                expected: Object::ARRAY(vec![
-                    Object::INTEGER(4),
-                    Object::INTEGER(5),
-                    Object::INTEGER(6),
-                ]),
-            },
-            VmTestCase {
-                input: r#"
-                    let x = fn() { 1 };
-                    let x = fn() { 2 };
-                    x()"#
-                    .to_string(),
-                expected: Object::INTEGER(2),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_shadowing_with_new_type() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let x = 4;
-                    let x = "string";
-                    x"#
-                .to_string(),
-                expected: Object::STRING("string".to_string()),
-            },
-            VmTestCase {
-                input: r#"
-                    let x = "string";
-                    let x = fn() { 1 };
-                    x()"#
-                    .to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"
-                    let x = fn() { 1 };
-                    let x = 5,
-                    x"#
-                .to_string(),
-                expected: Object::INTEGER(5),
-            },
-            VmTestCase {
-                input: r#"
-                    let x = 5;
-                    let x = [1,2,3];
-                    x"#
-                .to_string(),
-                expected: Object::ARRAY(vec![
-                    Object::INTEGER(1),
-                    Object::INTEGER(2),
-                    Object::INTEGER(3),
-                ]),
-            },
-            VmTestCase {
-                input: r#"
-                    let x = [1,2,3];
-                    let x = 5;
-                    x"#
-                .to_string(),
-                expected: Object::INTEGER(5),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_shadowing_using_previous_value() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    let b = a * a + 2
-                    b"#
-                .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    let a = a + 1;
-                    a"#
-                .to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = fn() { 
-                        let a = 1;
-                        a
-                    };
-                    a()
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"
-                    let f = fn(a){
-                        let a = 1;
-                        let a = a + 1;
-                        a
-                    };
-                    f(1)
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(2),
-            },
-            VmTestCase {
-                input: r#"
-                    let f = fn(){
-                        let a = 1;
-                        let h = fn(){
-                            let a = 2;
-                            a
-                        };
-                        h() + a
-                    };
-                    f()
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(3),
-            },
-            // Addition of a global variable a with 10 as its value
-            VmTestCase {
-                input: r#"
-                    let a = 10;
-                    let a = fn() { 
-                        let a = 1;
-                        a
-                    };
-                    a()
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 10;
-                    let f = fn(a){
-                        let a = 1;
-                        a
-                    };
-                    f(1) + a
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(11),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 10;
-                    let f = fn(a){
-                        let a = 1;
-                        let a = a + 1;
-                        a
-                    };
-                    f(1) + a
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(12),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 10;
-                    let f = fn(){
-                        let h = fn(){
-                            let a = 2;
-                            a
-                        };
-                        h()
-                    };
-                    f() + a
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(12),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 10;
-                    let f = fn(){
-                        let a = 1;
-                        let a = a + 1;
-                        let h = fn(){
-                            let a = 2;
-                            a
-                        };
-                        h() + a
-                    };
-                    f() + a
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(14),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_while_statements_without_break_or_continue() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    while (a < 100){
-                        let a = a + 1;
-                    }
-                    a
-                "#
-                .to_string(),
-                expected: Object::INTEGER(100),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    while (a < 0){
-                        let a = 100;
-                    }
-                    a
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(1),
-            },
-            VmTestCase {
-                input: r#"
-                    let a = 1;
-                    while(false){
-                        let a = 100;
-                    }
-                    a
-                    "#
-                .to_string(),
-                expected: Object::INTEGER(1),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_while_clean_up() {
-        // This tests makes sure that a while statement clears the stack correctly (which is
-        // different from the conditional behavior)
-        let tests = vec![VmTestCase {
-            input: r#"
-                    let a = 0;
-                    while (a < 10000){
-                        let a = a + 1;
-                        puts(a);
-                    }
-                    a
-                    "#
-            .to_string(),
-            expected: Object::INTEGER(10000),
-        }];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_break_from_while() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let a = 0;
-                    while (a < 10) {
-                        if (a == 5) {
-                            break;
-                        } 
-                        let a = a + 1;
-                    };
-                    a"#
-                .to_string(),
-                expected: Object::INTEGER(5),
-            },
-            VmTestCase {
-                input: r#"
-            let a = 0;
-            let c = 0;
-             while (a < 10) {
-                 let b = 0;
-                 while (b < 10) {
-                     if (b == 5) {
-                         break;
-                     }
-                     let b = b + 1;
-                     let c = c + 1;
-                 }
-                 let a = a + 1; 
-             };
-             c"#
-                .to_string(),
-                expected: Object::INTEGER(50),
-            },
-            VmTestCase {
-                input: r#"
-            let a = 0;
-            let c = 0;
-             while (a < 10) {
-                 if (a == 5) {
-                     break;
-                 }
-
-                 let b = 0;
-                 while (b < 10) {
-                     if (b == 5) {
-                         break;
-                     }
-                     let b = b + 1;
-                     let c = c + 1;
-                 }
-                 let a = a + 1; 
-             };
-             c"#
-                .to_string(),
-                expected: Object::INTEGER(25),
-            },
-            // The next tests will take care of the possible interference between the break and a function
-            VmTestCase {
-                input: r#"
-                let f = fn (a) {
-                    let c = 0;
-                    while (a < 10) {
-                        if (a == 5) {
-                            break;
-                        }
-
-                        let b = 0;
-                        while (b < 10) {
-                            if (b == 5) {
-                                break;
-                            }
-                            let b = b + 1;
-                            let c = c + 1;
-                        }
-                        let a = a + 1; 
-                    }
-                    c
-                };
-                f(0)"#
-                    .to_string(),
-                expected: Object::INTEGER(25),
-            },
-            VmTestCase {
-                input: r#"
-                let a = 0;
-                let c = 0;
-                while (a < 10) {
-                    if (a == 5) {
-                            break;
-                    }
-
-                    let f = fn () {
-                        let c = 0;
-                        let b = 0;
-                        while (b < 10) {
-                            if (b == 5) {
-                                break;
-                            }
-                            let b = b + 1;
-                            let c = c + 1;
-                        }
-                        c
-                    }
-
-                    let a = a + 1; 
-                    let c = c + f();
-                };
-                c"#
-                .to_string(),
-                expected: Object::INTEGER(25),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_continue_from_while() {
-        let tests = vec![
-            VmTestCase {
-                input: r#"
-                    let a = 0;
-                    let c = 0;
-                    while (a < 10) {
-                        let a = a + 1;
-                        if (a == 5) {
-                            let c  = c + 2;
-                            continue;
-                        } 
-                        let c = c + 1;
-                    };
-                    c"#
-                .to_string(),
-                expected: Object::INTEGER(11),
-            },
-            VmTestCase {
-                input: r#"
-            let a = 0;
-            let c = 0;
-             while (a < 10) {
-                 let b = 0;
-                 while (b < 10) {
-                     let b = b + 1;
-                     if (b == 5) {
-                         let c = c + 3;
-                         continue;
-                     }
-                     let c = c + 1;
-                 }
-                 let a = a + 1; 
-             };
-             c"#
-                .to_string(),
-                expected: Object::INTEGER(120),
-            },
-            // The next tests will take care of the possible interference between the continue and a function
-            VmTestCase {
-                input: r#"
-                let f = fn (a) {
-                    let c = 0;
-                    while (a < 10) {
-                        let b = 0;
-                        while (b < 10) {
-                            let b = b + 1;
-                            if (b == 5) {
-                                let c = c + 3;
-                                continue;
-                            }
-                            let c = c + 1;
-                        }
-                        let a = a + 1; 
-                    }
-                    c
-                };
-                f(0)"#
-                    .to_string(),
-                expected: Object::INTEGER(120),
-            },
-            VmTestCase {
-                input: r#"
-                let a = 0;
-                let c = 0;
-                while (a < 10) {
-                    let f = fn () {
-                        let c = 0;
-                        let b = 0;
-                        while (b < 10) {
-                            let b = b + 1;
-                            if (b == 5) {
-                                let c = c + 3;
-                                continue;
-                            }
-                            let c = c + 1;
-                        }
-                        c
-                    }
-
-                    let a = a + 1; 
-                    let c = c + f();
-                };
-                c"#
-                .to_string(),
-                expected: Object::INTEGER(120),
-            },
-        ];
-
-        run_vm_tests(tests);
-    }
-
-    #[test]
-    fn test_continue_and_break_in_while() {
-        let tests = vec![VmTestCase {
-            input: r#"
-                let a = 0;
-                let c = 0;
-                while (a < 10) {
-                    let a = a + 1; 
-                    if (a == 5) {
-                        let c = c + 3;
-                        continue;
-                    }
-                    if (a == 7) {
-                        break;
-                    }
-                    let c = c + 1;
-                }
-                c"#
-            .to_string(),
-            expected: Object::INTEGER(8),
-        }];
-        run_vm_tests(tests);
+        let mut vm = VM::new(bytecode);
+        vm.run().unwrap();
+        let got = vm.last_popped_stack_element().unwrap();
+        check_constants(&vec![test.expected], &vec![got]);
     }
 }
